patches are conditional, added 20 gauge
I attempted to make ammo only show up if it's parent exists in game, but
as far as I can tell the only way to do so is rely on whether mod
content exists for it instead of checking the ammo directly using CE
methods. I notice even CE cannot hide the EMP ammotypes as they are part
of the research tree and cannot be enabled/disabled selectively.