czar patch 2, added extra ammotypes
I added a couple new BP ammotypes, necessary for a more complete patch
for CZARsenal, as well as redesigned the way the patches will work.
Firearms will be seperated into <modname>_GunsPatch.xml, and ammotypes
will be in a <modname>_AmmoPatch.xml

This way, it should be easier to track down where a particular piece of
data comes from, as well as make it easier to implement updates when
needed.